Edit view livewire sales-stage sales-stage-lists

+++ resources/views/livewire/sales-stage/sales-stage-lists.blade.php
@@ -38,15 +38,17 @@
                 </div>
                 <div class="col-3 d-flex justify-content-center">
                     <div class="btn-group w-100" role="group">
-                        {{-- <button wire:click="$dispatch('add-sales-stage')" type="button"
-                            class="btn btn-primary disabled" data-toggle="modal" data-target="#modal-add-sales-stage">
-                            <i class="fas fa-plus"></i> Sales Stage
-                        </button> --}}
+
+                        @can('salesStage.create')
+                            <button wire:click="$dispatch('add-sales-stage')" type="button" class="btn btn-primary"
+                                data-toggle="modal" data-target="#modal-add-sales-stage">
+                                <i class="fas fa-plus"></i> Sales Stage
+                            </button>
+                        @endcan
+
                         <button wire:click="$dispatch('refresh-sales-stage')" type="button" class="btn btn-success">
                             <i class="fas fa-sync-alt"></i> Refresh
                         </button>
-
-                        @livewire('sales-stage.sales-stage-create')
                     </div>
                 </div>
             </div>
@@ -67,14 +69,18 @@
         <!-- /.card-header -->
         <div class="card-body">
             <div class="table-responsive">
-                <table class="table table-hover">
-                    <thead>
+                <table class="table table-sm table-hover table-bordered">
+                    <thead class="table-info">
                         <th scope="col">#</th>
                         <th scope="col">Created</th>
                         <th scope="col">Updated</th>
                         <th scope="col">Id</th>
                         <th scope="col">Sales Stage Name</th>
-                        <th scope="col" style="width: 100px">Action</th>
+                        @can('salesStage.edit')
+                            <th scope="col" colspan="2">Action</th>
+                        @elsecan('salesStage.delete')
+                            <th scope="col" colspan="2">Action</th>
+                        @endcan
                     </thead>
 
                     <tbody>
@@ -119,23 +125,34 @@
                                 </td>
                                 <td>{{ $item->id }}</td>
                                 <td>{{ $item->name }}</td>
-                                <td>
-                                    {{-- <button wire:click.prevent="$dispatch('edit-sales-stage',{id:{{ $item->id }}})"
-                                        type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
-                                        data-target="#modal-edit-sales-stage">
-                                        <i class="fas fa-edit"></i>
-                                    </button>
-                                    <button
-                                        wire:click.prevent="deleteSalesStage({{ $item->id }},{{ "'" . str_replace("'", '', $item->name) . "'" }})"
-                                        class="btn btn-danger btn-sm">
-                                        <i class="fas fa-trash"></i>
-                                    </button> --}}
-                                </td>
+                                @can('salesStage.edit')
+                                    <td style="width: 45px" class="p-1 text-center">
+                                        <button wire:click.prevent="$dispatch('edit-sales-stage',{id:{{ $item->id }}})"
+                                            type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
+                                            data-target="#modal-edit-sales-stage">
+                                            <i class="fas fa-edit"></i>
+                                        </button>
+                                    </td>
+                                @endcan
+
+                                @can('salesStage.delete')
+                                    <td style="width: 45px" class="p-1 text-center">
+                                        <button
+                                            wire:click.prevent="deleteSalesStage({{ $item->id }},{{ "'" . str_replace("'", '', $item->name) . "'" }})"
+                                            class="btn btn-danger btn-sm">
+                                            <i class="fas fa-trash"></i>
+                                        </button>
+                                    </td>
+                                @endcan
                             </tr>
                         @endforeach
-                        @livewire('sales-stage.sales-stage-edit')
                     </tbody>
                 </table>
+
+                @livewire('sales-stage.sales-stage-create')
+
+                @livewire('sales-stage.sales-stage-edit')
+
             </div>
         </div>
         <!-- /.card-body -->
